+++ src/TMApplication/ViewModels/DivisionViewModel.cs
@@ -14,15 +14,15 @@ public record DivisionViewModel(
 
 public record DivisionPlayerViewModel(
     string Name,
-    decimal TotalPoints = 0,
-    ushort Wins = 0,
-    ushort Cla = 0,
-    ushort Supplies = 0,
-    ushort PowerTokens = 0,
+    double TotalPoints = 0,
+    int Wins = 0,
+    int Cla = 0,
+    int Supplies = 0,
+    int PowerTokens = 0,
     double MinutesPerMove = 0,
-    ushort Moves = 0,
+    int Moves = 0,
     PlayerHouseViewModel[]? Houses = null,
-    decimal TotalPenaltyPoints = 0,
+    double TotalPenaltyPoints = 0,
     PlayerPenaltyViewModel[]? Penalties = null)
 {
     public PlayerHouseViewModel GetHouse(House house) =>
@@ -31,20 +31,20 @@ public record DivisionPlayerViewModel(
 }
 
 public record PlayerHouseViewModel(
-    uint? Game,
+    int? Game,
     House House,
     bool IsWinner = false,
-    decimal Points = 0,
-    ushort BattlePenalty = 0,
-    ushort Strongholds = 0,
-    ushort Castles = 0,
-    ushort Cla = 0,
-    ushort Supplies = 0,
-    ushort PowerTokens = 0,
+    double Points = 0,
+    int BattlePenalty = 0,
+    int Strongholds = 0,
+    int Castles = 0,
+    int Cla = 0,
+    int Supplies = 0,
+    int PowerTokens = 0,
     double MinutesPerMove = 0,
-    ushort Moves = 0);
+    int Moves = 0);
 
 public record PlayerPenaltyViewModel(
-    uint? Game,
-    decimal Points,
+    int? Game,
+    double Points,
     string Details);
