+++ src/TMGameImporter/Services/Import/DivisionImportingService.cs
@@ -68,15 +68,15 @@ internal class DivisionImportingService
         var houseResults = GetHouses(playerName, division.Replacements ?? Array.Empty<Replacement>(), games, scoring);
         var penalties = GetPenalties(playerName, division, houseResults);
         var penaltiesPoints = penalties.Sum(penalty => penalty.Points);
-        var moves = (ushort)houseResults.Sum(houseResult => houseResult.Moves);
+        var moves = houseResults.Sum(houseResult => houseResult.Moves);
         var minutesPerMove = moves == 0 ?
             0 : houseResults.Sum(houseResult => houseResult.MinutesPerMove * houseResult.Moves) / moves;
         return new PlayerResult(playerName,
             houseResults.Sum(houseResult => houseResult.Points) - penaltiesPoints,
-            (ushort)houseResults.Count(houseResult => houseResult.IsWinner),
-            (ushort)houseResults.Sum(houseResult => houseResult.Cla),
-            (ushort)houseResults.Sum(houseResult => houseResult.Supplies),
-            (ushort)houseResults.Sum(houseResult => houseResult.PowerTokens),
+            houseResults.Count(houseResult => houseResult.IsWinner),
+            houseResults.Sum(houseResult => houseResult.Cla),
+            houseResults.Sum(houseResult => houseResult.Supplies),
+            houseResults.Sum(houseResult => houseResult.PowerTokens),
             minutesPerMove,
             moves,
             houseResults,
@@ -113,18 +113,18 @@ internal class DivisionImportingService
             houseScore.Moves);
     }
 
-    private static decimal GetPointsForGame(HouseScore houseScore, bool isWinner, Scoring scoring) =>
+    private static double GetPointsForGame(HouseScore houseScore, bool isWinner, Scoring scoring) =>
         scoring.PointsPerStronghold * houseScore.Strongholds +
         scoring.PointsPerCastle * houseScore.Castles +
         (isWinner ? scoring.PointsPerWin : 0);
 
-    private static ushort GetBattlePenalty(Game game, HouseScore houseScore, Scoring scoring)
+    private static int GetBattlePenalty(Game game, HouseScore houseScore, Scoring scoring)
     {
         if (game.Turn < 10)
             return 0;
 
         var battlesBefore10ThTurn = houseScore.BattlesInTurn[..9].Sum(battlesInTurn => battlesInTurn);
-        return (ushort)Math.Max(0, scoring.RequiredBattlesBefore10thTurn - battlesBefore10ThTurn);
+        return Math.Max(0, scoring.RequiredBattlesBefore10thTurn - battlesBefore10ThTurn);
     }
 
     private static PlayerPenalty[] GetPenalties(string playerName, Division division, IEnumerable<HouseResult> houseResults)
